AmountValueIcon component
